# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Freight bidding workflow orchestrator",
        "description": "Automated competitive freight bidding API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/quotes": {
            "get": {
                "tags": [
                    "Freight Quotes"
                ],
                "summary": "Get Quotes",
                "operationId": "get_quotes_api_quotes_get",
                "parameters": [
                    {
                        "name": "X-User-Email",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-User-Email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FreightQuoteResponse"
                                    },
                                    "title": "Response Get Quotes Api Quotes Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Freight Quotes"
                ],
                "summary": "Create Quote",
                "operationId": "create_quote_api_quotes_post",
                "parameters": [
                    {
                        "name": "X-User-Email",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-User-Email"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FreightQuoteCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FreightQuoteResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quotes/{quote_id}": {
            "get": {
                "tags": [
                    "Freight Quotes"
                ],
                "summary": "Get Quote",
                "operationId": "get_quote_api_quotes__quote_id__get",
                "parameters": [
                    {
                        "name": "quote_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Quote Id"
                        }
                    },
                    {
                        "name": "X-User-Email",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-User-Email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FreightQuoteResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarrierBidResponse": {
                "properties": {
                    "carrier_id": {
                        "type": "integer",
                        "title": "Carrier Id"
                    },
                    "round": {
                        "type": "integer",
                        "title": "Round",
                        "default": 1
                    },
                    "bid_amount": {
                        "type": "number",
                        "title": "Bid Amount"
                    },
                    "transit_time_days": {
                        "type": "integer",
                        "title": "Transit Time Days"
                    },
                    "pickup_window": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pickup Window",
                        "default": "09:00 - 17:00"
                    },
                    "accessorials_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Accessorials Text",
                        "default": ""
                    },
                    "service_level": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Service Level",
                        "default": "Standard LTL"
                    },
                    "notes": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notes",
                        "default": ""
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "freight_quote_id": {
                        "type": "string",
                        "title": "Freight Quote Id"
                    },
                    "received_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Received At"
                    },
                    "is_winning": {
                        "type": "boolean",
                        "title": "Is Winning"
                    }
                },
                "type": "object",
                "required": [
                    "carrier_id",
                    "bid_amount",
                    "transit_time_days",
                    "id",
                    "freight_quote_id",
                    "received_at",
                    "is_winning"
                ],
                "title": "CarrierBidResponse"
            },
...
            "CarrierResponse": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "competitiveness_score": {
                        "type": "number",
                        "title": "Competitiveness Score",
                        "default": 0.0
                    },
                    "is_override": {
                        "type": "boolean",
                        "title": "Is Override",
                        "default": false
                    },
                    "simulated_score": {
                        "type": "number",
                        "title": "Simulated Score",
                        "default": 0.0
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "calculated_competitiveness_score": {
                        "type": "number",
                        "title": "Calculated Competitiveness Score",
                        "default": 0.0
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "email",
                    "id"
                ],
                "title": "CarrierResponse"
            },
...
            "CustomerResponse": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "default_markup_percent": {
                        "type": "number",
                        "title": "Default Markup Percent",
                        "default": 10.0
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "email",
                    "id",
                    "created_at"
                ],
                "title": "CustomerResponse"
            },
...
            "FreightQuoteCreate": {
                "properties": {
                    "customer_id": {
                        "type": "integer",
                        "title": "Customer Id"
                    },
                    "origin": {
                        "type": "string",
                        "title": "Origin"
                    },
                    "destination": {
                        "type": "string",
                        "title": "Destination"
                    },
                    "weight_lbs": {
                        "type": "number",
                        "title": "Weight Lbs"
                    },
                    "dimensions": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Dimensions",
                        "default": "48x48x48"
                    },
                    "freight_class": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Freight Class",
                        "default": "70"
                    },
                    "hazmat": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hazmat",
                        "default": false
                    },
                    "accessorials": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Accessorials",
                        "default": ""
                    },
                    "pickup_date": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Pickup Date"
                    }
                },
                "type": "object",
                "required": [
                    "customer_id",
                    "origin",
                    "destination",
                    "weight_lbs",
                    "pickup_date"
                ],
                "title": "FreightQuoteCreate"
            },
            "FreightQuoteResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "customer_id": {
                        "type": "integer",
                        "title": "Customer Id"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerResponse"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "origin": {
                        "type": "string",
                        "title": "Origin"
                    },
                    "destination": {
                        "type": "string",
                        "title": "Destination"
                    },
                    "weight_lbs": {
                        "type": "number",
                        "title": "Weight Lbs"
                    },
                    "dimensions": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Dimensions"
                    },
                    "freight_class": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Freight Class"
                    },
                    "hazmat": {
                        "type": "boolean",
                        "title": "Hazmat"
                    },
                    "accessorials": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Accessorials"
                    },
                    "pickup_date": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Pickup Date"
                    },
                    "first_round_ends_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "First Round Ends At"
                    },
                    "rebid_round_ends_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rebid Round Ends At"
                    },
                    "quote_expires_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Quote Expires At"
                    },
                    "cost_price": {
                        "type": "number",
                        "title": "Cost Price"
                    },
                    "sell_price": {
                        "type": "number",
                        "title": "Sell Price"
                    },
                    "markup_percent": {
                        "type": "number",
                        "title": "Markup Percent"
                    },
                    "margin_amt": {
                        "type": "number",
                        "title": "Margin Amt"
                    },
                    "margin_pct": {
                        "type": "number",
                        "title": "Margin Pct"
                    },
                    "winning_carrier_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Winning Carrier Id"
                    },
                    "winning_carrier": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/CarrierResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "lost_reason": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Lost Reason"
                    },
                    "competitor_info": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Competitor Info"
                    },
                    "payment_status": {
                        "type": "string",
                        "title": "Payment Status"
                    },
                    "bol_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bol Url"
                    },
                    "invoice_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Invoice Url"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "bids": {
                        "items": {
                            "$ref": "#/components/schemas/CarrierBidResponse"
                        },
                        "type": "array",
                        "title": "Bids",
                        "default": []
                    },
                    "transitions": {
                        "items": {
                            "$ref": "#/components/schemas/StateTransitionResponse"
                        },
                        "type": "array",
                        "title": "Transitions",
                        "default": []
                    },
                    "rfqs": {
                        "items": {
                            "$ref": "#/components/schemas/RequestForQuoteResponse"
                        },
                        "type": "array",
                        "title": "Rfqs",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "customer_id",
                    "customer",
                    "status",
                    "origin",
                    "destination",
                    "weight_lbs",
                    "dimensions",
                    "freight_class",
                    "hazmat",
                    "accessorials",
                    "pickup_date",
                    "first_round_ends_at",
                    "rebid_round_ends_at",
                    "quote_expires_at",
                    "cost_price",
                    "sell_price",
                    "markup_percent",
                    "margin_amt",
                    "margin_pct",
                    "winning_carrier_id",
                    "winning_carrier",
                    "lost_reason",
                    "competitor_info",
                    "payment_status",
                    "bol_url",
                    "invoice_url",
                    "created_at",
                    "updated_at"
                ],
                "title": "FreightQuoteResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RequestForQuoteResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "freight_quote_id": {
                        "type": "string",
                        "title": "Freight Quote Id"
                    },
                    "carrier_id": {
                        "type": "integer",
                        "title": "Carrier Id"
                    },
                    "supplier_id": {
                        "type": "integer",
                        "title": "Supplier Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "sent_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Sent At"
                    },
                    "subject": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Subject"
                    },
                    "body": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Body"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "freight_quote_id",
                    "carrier_id",
                    "supplier_id",
                    "status",
                    "sent_at"
                ],
                "title": "RequestForQuoteResponse"
            },
...
            "StateTransitionResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "from_status": {
                        "type": "string",
                        "title": "From Status"
                    },
                    "to_status": {
                        "type": "string",
                        "title": "To Status"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Timestamp"
                    },
                    "notes": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notes"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "from_status",
                    "to_status",
                    "timestamp"
                ],
                "title": "StateTransitionResponse"
            },
...
        }
    }
}